Fixed lint problem. Refactored ./tests folder.
